% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_current_gwascat.R
\name{as_GRanges}
\alias{as_GRanges}
\title{produce a GRanges from gwascat tibble}
\usage{
as_GRanges(
  x,
  fixup = TRUE,
  short = TRUE,
  for_short = c("PUBMEDID", "DATE", "DISEASE/TRAIT", "SNPS"),
  genome_tag = "GRCh38"
)
}
\arguments{
\item{x}{a tibble from `get_cached_gwascat()`}

\item{fixup}{logical(1) defaults to TRUE: to deal with missing `CHR_ID` and
other unexpected records seen in March 2025, apply fixsnps function before
transforming; failure to do this may lead to errors in GRanges production}

\item{short}{logical(1) if TRUE only keep selected columns in mcols}

\item{for_short}{character() column names to keep in mcols}

\item{genome_tag}{character(1) defaults to "GRCh38"}
}
\description{
produce a GRanges from gwascat tibble
}
