% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{chklocs}
\alias{chklocs}
\title{return TRUE if all named SNPs with locations in both
the SNPlocs package and the gwascat agree}
\usage{
chklocs(chrtag = "20", gwwl = gwrngs19)
}
\arguments{
\item{chrtag}{character, chromosome identifier}

\item{gwwl}{instance of \code{{gwaswloc}}}
}
\description{
return TRUE if all named SNPs with locations in both
the SNPlocs package and the gwascat agree
}
