% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwascat-package.R
\docType{data}
\name{ebicat37}
\alias{ebicat37}
\title{a legacy representation of EBI's GWAS catalog with loci annotated to positions on GRCh37}
\format{
gwaswloc instance
}
\usage{
ebicat37
}
\description{
a legacy representation of EBI's GWAS catalog with loci annotated to positions on GRCh37
}
\keyword{datasets}
