% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwascat-package.R
\docType{data}
\name{ebicat37UCSC}
\alias{ebicat37UCSC}
\title{a legacy representation of EBI's GWAS catalog with loci annotated to positions on GRCh37 but
UCSC chromosome names}
\format{An object of class \code{gwaswloc} of length 22688.}
\usage{
ebicat37UCSC
}
\description{
a legacy representation of EBI's GWAS catalog with loci annotated to positions on GRCh37 but
UCSC chromosome names
}
\keyword{datasets}
