% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwascat-package.R
\docType{data}
\name{ebicat_b38}
\alias{ebicat_b38}
\title{a legacy representation of EBI's GWAS catalog with loci annotated to positions on GRCh38}
\format{
gwaswloc instance
}
\usage{
ebicat_b38
}
\description{
a legacy representation of EBI's GWAS catalog with loci annotated to positions on GRCh38
}
\keyword{datasets}
