% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixsnps.R
\name{fixsnps}
\alias{fixsnps}
\title{this function looks for clues to resolve records that have missing CHR_ID or CHR_POS,
for which address info can be inferred from SNPS field}
\usage{
fixsnps(x)
}
\arguments{
\item{x}{output of makeCurrentGwascat or get_cached_gwascat}
}
\description{
this function looks for clues to resolve records that have missing CHR_ID or CHR_POS,
for which address info can be inferred from SNPS field
}
