% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{getRsids,gwaswloc-method}
\alias{getRsids,gwaswloc-method}
\title{specific snp name retrieval}
\usage{
\S4method{getRsids}{gwaswloc}(x)
}
\arguments{
\item{x}{gwaswloc}
}
\description{
specific snp name retrieval
}
