% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{getTraits,gwaswloc-method}
\alias{getTraits,gwaswloc-method}
\title{specific trait retrieval}
\usage{
\S4method{getTraits}{gwaswloc}(x)
}
\arguments{
\item{x}{gwaswloc}
}
\description{
specific trait retrieval
}
