% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gg17N}
\alias{gg17N}
\title{genotype matrix from chr17 1000 genomes}
\format{
matrix
}
\usage{
data(gg17N)
}
\description{
genotype matrix from chr17 1000 genomes
}
\examples{
data(gg17N)
gg17N[1:4,1:4]
}
\keyword{datasets}
