% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gr6.0_hg38}
\alias{gr6.0_hg38}
\title{image of locon6 in GRanges, lifted over to hg38}
\format{
GRanges instance
}
\usage{
data(gr6.0_hg38)
}
\description{
image of locon6 in GRanges, lifted over to hg38
}
\keyword{datasets}
