% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwastagger}
\alias{gwastagger}
\title{GRanges with LD information on 9998 SNP}
\format{
GRanges
}
\usage{
data(gwastagger)
}
\description{
GRanges with LD information on 9998 SNP
}
\keyword{datasets}
