% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwcat_snapshot.R
\name{gwcat_snapshot}
\alias{gwcat_snapshot}
\title{use AnnotationHub snapshot as basis for gwaswloc structure creation}
\usage{
gwcat_snapshot(x, fixNonASCII = TRUE)
}
\arguments{
\item{x}{inherits from data.frame, with columns consistent with EBI table}

\item{fixNonASCII}{logical(1) if TRUE, use iconv to replace non-ASCII character, important
for CMD check but perhaps not important for applied use}
}
\description{
use AnnotationHub snapshot as basis for gwaswloc structure creation
}
\examples{
ah = AnnotationHub::AnnotationHub()
entitytab = AnnotationHub::query(ah, "gwascatData")
cand = names(entitytab)[1]
stopifnot(nchar(cand)>0)
tab = ah[[cand]]
gww = gwcat_snapshot(tab)
gww
length(gww)
}
