% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{locs4trait}
\alias{locs4trait}
\title{get locations for SNP affecting a selected trait}
\usage{
locs4trait(gwwl, trait, tag = "DISEASE/TRAIT")
}
\arguments{
\item{gwwl}{instance of \code{\linkS4class{gwaswloc}}}

\item{trait}{character, name of trait}

\item{tag}{character, name of field to be used for trait enumeration}
}
\description{
get locations for SNP affecting a selected trait
}
