% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gwas_dataframe.R
\name{process_gwas_dataframe}
\alias{process_gwas_dataframe}
\title{convert GWAS catalog data.frame to gwaswloc, a GRanges extension with simple show method}
\usage{
process_gwas_dataframe(df)
}
\arguments{
\item{df}{data.frame}
}
\description{
convert GWAS catalog data.frame to gwaswloc, a GRanges extension with simple show method
}
