% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{[,gwaswloc,ANY,ANY,ANY-method}
\alias{[,gwaswloc,ANY,ANY,ANY-method}
\title{extractor for gwaswloc}
\usage{
\S4method{[}{gwaswloc,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{gwaswloc}

\item{i}{index}

\item{j}{index}

\item{\dots}{addtl indices}

\item{drop}{logical(1)}
}
\description{
extractor for gwaswloc
}
