% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{subsetByChromosome,gwaswloc-method}
\alias{subsetByChromosome,gwaswloc-method}
\title{specific trait subsetting}
\usage{
\S4method{subsetByChromosome}{gwaswloc}(x, ch)
}
\arguments{
\item{x}{gwaswloc}

\item{ch}{character vector of chromosomes}
}
\description{
specific trait subsetting
}
