% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{subsetByChromosome}
\alias{subsetByChromosome}
\title{generic trait subsetting}
\usage{
subsetByChromosome(x, ch)
}
\arguments{
\item{x}{gwaswloc}

\item{ch}{character vector of chromosomes}
}
\description{
generic trait subsetting
}
