% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{subsetByTraits,gwaswloc-method}
\alias{subsetByTraits,gwaswloc-method}
\title{specific trait subsetting}
\usage{
\S4method{subsetByTraits}{gwaswloc}(x, tr)
}
\arguments{
\item{x}{gwaswloc}

\item{tr}{character vector of traits}
}
\description{
specific trait subsetting
}
