% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{subsetByTraits}
\alias{subsetByTraits}
\title{generic trait subsetting}
\usage{
subsetByTraits(x, tr)
}
\arguments{
\item{x}{gwaswloc}

\item{tr}{character vector of traits}
}
\description{
generic trait subsetting
}
