% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{topTraits}
\alias{topTraits}
\title{operations on GWAS catalog}
\usage{
topTraits(gwwl, n = 10, tag = "DISEASE/TRAIT")
}
\arguments{
\item{gwwl}{instance of \code{{gwaswloc}}}

\item{n}{numeric, number of traits to report}

\item{tag}{character, name of field to be used for trait enumeration}
}
\value{
\code{topTraits} returns a character vector of most frequently
occurring traits in the database

\code{locs4trait} returns a \code{{gwaswloc}} object with
records defining associations to the specified trait

\code{chklocs} returns a logical that is TRUE when the asserted locations of
SNP in the GWAS catalog agree with the locations given in the dbSNP package
SNPlocs.Hsapiens.dbSNP144.GRCh37
}
\description{
operations on GWAS catalog
}
\examples{

data(ebicat_2020_04_30)
topTraits(ebicat_2020_04_30)

}
\author{
VJ Carey <stvjc@channing.harvard.edu>
}
\keyword{models}
