% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ebicat_2020_04_30}
\alias{ebicat_2020_04_30}
\title{serialized gwaswloc instance from april 30 2020}
\format{
gwaswloc instance
}
\usage{
ebicat_2020_04_30
}
\description{
serialized gwaswloc instance from april 30 2020
}
\keyword{datasets}
