% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_current_gwascat.R
\name{get_cached_gwascat}
\alias{get_cached_gwascat}
\title{use BiocFileCache to retrieve and keep an image of the tsv file}
\usage{
get_cached_gwascat(
  url = "http://www.ebi.ac.uk/gwas/api/search/downloads/alternative",
  cache = BiocFileCache::BiocFileCache(),
  ...
)
}
\arguments{
\item{url}{character(1) url to use}

\item{cache}{BiocFileCache::BiocFileCache instance}

\item{\dots}{passed to bfcadd}
}
\description{
use BiocFileCache to retrieve and keep an image of the tsv file
}
\note{
will If query of cache with 'ebi.ac.uk/gwas' returns 0-row tibble,
will populate cache with bfcadd.  Uses readr::read_tsv on cache content to return tibble.
The etag field does not seem to be used at EBI, thus user must check for updates.
}
