% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gw6.rs_17}
\alias{gw6.rs_17}
\title{character vector of rs numbers for SNP on chr17}
\format{
character vector
}
\usage{
gw6.rs_17
}
\description{
character vector of rs numbers for SNP on chr17
}
\keyword{datasets}
