% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwascat-package.R
\docType{class}
\name{gwaswloc-class}
\alias{gwaswloc-class}
\alias{[,gwaswloc,ANY,ANY,ANY-method}
\alias{[,gwaswloc,ANY-method}
\alias{[,gwaswloc-method}
\alias{show,gwaswloc-method}
\alias{getRsids,gwaswloc-method}
\alias{getTraits,gwaswloc-method}
\alias{subsetByChromosome,gwaswloc-method}
\alias{subsetByTraits,gwaswloc-method}
\alias{getTraits}
\alias{getRsids}
\alias{subsetByChromosome}
\alias{subsetByTraits}
\title{Class \code{"gwaswloc"}}
\description{
A container for GRanges instances representing information in the NHGRI GWAS
catalog.
}
\note{
In gwascat package 1.9.6 and earlier, the globally accessible
\code{gwaswloc} instance \code{gwrngs} was created upon attachment.  This is
no longer the case.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("gwaswloc", ...)}. Any GRanges instance can be supplied.
}

\examples{

showClass("gwaswloc")

}
\references{
%% ~~put references to the literature/web site here~~
\url{http://www.ebi.ac.uk/gwas/}
}
\author{
VJ Carey <stvjc@channing.harvard.edu>
}
\keyword{classes}
