% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwascat-package.R
\docType{data}
\name{gwdf_2012_02_02}
\alias{gwdf_2012_02_02}
\alias{gwdf_2014_09_08}
\title{internal data frame for NHGRI GWAS catalog}
\format{A data frame with 17832 observations on the following 34 variables.}
\source{
\url{http://www.ebi.ac.uk/gwas/}
}
\value{
a DataFrame with (character) columns
"Date Added to Catalog",
"PUBMEDID",
"First Author",
"Date",
"Journal",
"Link",
"Study",
"Disease/Trait",
"Initial Sample Size",
"Replication Sample Size",
"Region",
"Chr_id",
"Chr_pos",
"Reported Gene(s)",
"Mapped_gene",
"Upstream_gene_id",
"Downstream_gene_id",
"Snp_gene_ids",
"Upstream_gene_distance",
"Downstream_gene_distance",
"Strongest SNP-Risk Allele",
"SNPs",
"Merged",
"Snp_id_current",
"Context",
"Intergenic",
"Risk Allele Frequency",
"p-Value",
"Pvalue_mlog",
"p-Value (text)",
"OR or beta",
"95% CI (text)",
"Platform \[SNPs passing QC\]",
"CNV"
}
\description{
convenience container for imported table from NHGRI GWAS catalog
}
\note{
In versions prior to 1.9.6, The .onAttach function specifies which
data frame is transformed to GRanges.  This is now managed manually.
}
\examples{

\dontrun{
data(gwdf_2014_09_08)
# try gwascat:::gwdf2GRanges on this data.frame
}

}
\keyword{datasets}
