% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwascat-package.R
\docType{data}
\name{locon6}
\alias{locon6}
\title{location information for 10000 SNPs probed on Affy GW 6.0}
\format{A data frame with 10000 observations on the following 3 variables.
\describe{ \item{dbsnp_rs_id}{a character vector}
\item{chrom}{a character vector} \item{physical_pos}{a
numeric vector} }}
\description{
location information for 10000 SNPs probed on Affy GW 6.0
}
\details{
extracted from pd.genomewidesnp.6 v 1.4.0; for demonstration purposes
}
\examples{

data(locon6)
str(locon6) 

}
\keyword{datasets}
